requirements.txt
argcomplete==1.10.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
docx2txt==0.8
EbookLib==0.17.1
entrypoints==0.3
extract-msg==0.23.1
Flask==1.1.1
Flask-Cors==3.0.8
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.24.3
idna==2.8
IMAPClient==2.1.0
img2pdf==0.3.3
importlib-metadata==0.23
imutils==0.5.3
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
lxml==4.4.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numpy==1.17.3
olefile==0.46
opencv-python==4.1.1.26
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pdfminer==20191020
pdfminer.six==20181108
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.9.0
Pygments==2.4.2
PyPDF2==1.26.0
pyrsistent==0.15.4
pytesseract==0.3.0
python-dateutil==2.8.0
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
#pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
terminado==0.8.2
#tesseract==0.1.3
#tesserocr==2.4.0
testpath==0.4.2
textract==1.6.3
tika==1.19
tornado==6.0.3
traitlets==4.3.3
tzlocal==1.5.1
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.2.2
zipp==0.6.0
flask
flask-cors
